Move NKFile to metadata conversion methods to a new extension for NextcloudItemMetadataTable, where they are better placed
